+++ lib/qonfig.rb
@@ -23,4 +23,12 @@ module Qonfig
   require_relative 'qonfig/data_set'
   require_relative 'qonfig/data_set/class_builder'
   require_relative 'qonfig/configurable'
+  require_relative 'qonfig/plugins/registry'
+  require_relative 'qonfig/plugins'
+  require_relative 'qonfig/plugins/access_mixin'
+  require_relative 'qonfig/plugins/abstract'
+
+  # @api public
+  # @since 0.4.0
+  extend Plugins::AccessMixin
 end