+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2018-06-24
+- Introduce Plugin Ecosystem (`Qonfig::Plugins`):
+  - load plugin: `Qonfig.plugin('plugin_name')` or `Qonfig.plugin(:plugin_name)`;
+  - get registered plugins: `Qonfig.plugins #=> array of strings`
+
 ## [0.3.0] - 2018-06-13
 ### Added
 - Improved configuration process: `#configure` can take a hash as a configuration `[option key => option]`